feat: ➕ add fetchTournamentTree
